Makefile to create epub from markdown
The makefile uses pandoc to generate epub,
and epubcheck to validate the generated epub.

All images are put into one global images folder.
Images from different chapters with the same names should
be renamed to avoid collisions.

Chapter names are in single variable, which is used
to copy images and produce the epub.

Makefile results in target folder.

Target folder added to .gitignore.

Epub validation is an optional check.