Added cover image to cover in pdf.
Added css file used for epub generation.

Currently ensuring captions are displayed in italics,
to make it clear they are different from the main text.

Fixed target to inspect epub.

Fixed vagrant provisioning so that proper ruby is loaded.

Jekyll needs ruby 2 -- the code was already in the provision file,
but in comments, for some reason. Uncommented it.

epub-style messes up table of contents. Fixed.